Fix for copying menu's from subsite to main site
When copying menu's to the main site, the switch_to_blog function showed erratic behaviour and caused the update_option function to perform updates on the wrong blog id. This fixes that by directly updating the meta data in the options table. It may not be the prettiest solution, but it is a reliable one.